webservices/requirements.txt
Django==5.1.1
psycopg2-binary==2.9.9
djangorestframework==3.15.2
pillow==11.0.0
gunicorn==20.1.0
google-cloud-secret-manager==2.16.2
google-api-core==2.11.1
google-auth==2.16.3
google-cloud-core==2.3.3
proto-plus==1.22.2
grpcio==1.56.0
grpcio-status==1.56.0
protobuf==4.21.12
django-storages[google]==1.13